Extract DataGenerator static helpers into src/Profiles/Data/
Closes #182
