+++ src/Profiles/DataGenerator.cs
@@ -410,223 +410,3 @@ internal class DataGenerator
         return new string(buffer);
     }
 }
-
-/// <summary>
-/// Lorem ipsum text generator.
-/// </summary>
-internal static class LoremIpsum
-{
-    private static readonly string[] Words =
-    {
-        "lorem", "ipsum", "dolor", "sit", "amet", "consectetur", "adipiscing", "elit",
-        "sed", "do", "eiusmod", "tempor", "incididunt", "ut", "labore", "et", "dolore",
-        "magna", "aliqua", "enim", "ad", "minim", "veniam", "quis", "nostrud",
-        "exercitation", "ullamco", "laboris", "nisi", "aliquip", "ex", "ea", "commodo",
-        "consequat", "duis", "aute", "irure", "in", "reprehenderit", "voluptate",
-        "velit", "esse", "cillum", "fugiat", "nulla", "pariatur", "excepteur", "sint",
-        "occaecat", "cupidatat", "non", "proident", "sunt", "culpa", "qui", "officia",
-        "deserunt", "mollit", "anim", "id", "est", "laborum",
-    };
-
-    /// <summary>
-    /// Gets a paragraph of Lorem ipsum text.
-    /// </summary>
-    /// <param name="random">Random instance.</param>
-    /// <returns>A paragraph.</returns>
-    public static string GetParagraph(Random random)
-    {
-        var sentenceCount = random.Next(3, 7);
-        var sb = new StringBuilder();
-
-        for (int i = 0; i < sentenceCount; i++)
-        {
-            if (i > 0)
-            {
-                sb.Append(' ');
-            }
-
-            AppendSentence(sb, random);
-        }
-
-        return sb.ToString();
-    }
-
-    /// <summary>
-    /// Gets a sentence of Lorem ipsum text.
-    /// </summary>
-    /// <param name="random">Random instance.</param>
-    /// <returns>A sentence.</returns>
-    public static string GetSentence(Random random)
-    {
-        var sb = new StringBuilder();
-        AppendSentence(sb, random);
-        return sb.ToString();
-    }
-
-    private static void AppendSentence(StringBuilder sb, Random random)
-    {
-        var wordCount = random.Next(8, 20);
-        for (int i = 0; i < wordCount; i++)
-        {
-            if (i > 0)
-            {
-                sb.Append(' ');
-            }
-
-            var word = Words[random.Next(Words.Length)];
-            if (i == 0)
-            {
-                sb.Append(char.ToUpperInvariant(word[0]));
-                if (word.Length > 1)
-                {
-                    sb.Append(word, 1, word.Length - 1);
-                }
-            }
-            else
-            {
-                sb.Append(word);
-            }
-        }
-
-        sb.Append('.');
-    }
-}
-
-/// <summary>
-/// Name data for generation.
-/// </summary>
-internal static class Names
-{
-    /// <summary>
-    /// First names for generation.
-    /// </summary>
-    public static readonly string[] FirstNames =
-    {
-        "James", "Mary", "John", "Patricia", "Robert", "Jennifer", "Michael", "Linda",
-        "William", "Elizabeth", "David", "Barbara", "Richard", "Susan", "Joseph", "Jessica",
-        "Thomas", "Sarah", "Charles", "Karen", "Christopher", "Nancy", "Daniel", "Lisa",
-        "Matthew", "Betty", "Anthony", "Margaret", "Mark", "Sandra", "Donald", "Ashley",
-    };
-
-    /// <summary>
-    /// Lowercased first names for generation.
-    /// </summary>
-    internal static readonly string[] FirstNamesLower;
-
-    /// <summary>
-    /// Last names for generation.
-    /// </summary>
-    public static readonly string[] LastNames =
-    {
-        "Smith", "Johnson", "Williams", "Brown", "Jones", "Garcia", "Miller", "Davis",
-        "Rodriguez", "Martinez", "Hernandez", "Lopez", "Gonzalez", "Wilson", "Anderson",
-        "Thomas", "Taylor", "Moore", "Jackson", "Martin", "Lee", "Perez", "Thompson",
-        "White", "Harris", "Sanchez", "Clark", "Ramirez", "Lewis", "Robinson", "Walker",
-    };
-
-    /// <summary>
-    /// Lowercased last names for generation.
-    /// </summary>
-    internal static readonly string[] LastNamesLower;
-
-    /// <summary>
-    /// Initializes static members of the <see cref="Names"/> class.
-    /// </summary>
-    static Names()
-    {
-        FirstNamesLower = FirstNames.Select(n => n.ToLowerInvariant()).ToArray();
-        LastNamesLower = LastNames.Select(n => n.ToLowerInvariant()).ToArray();
-    }
-}
-
-/// <summary>
-/// Email subject generator.
-/// </summary>
-internal static class EmailSubjects
-{
-    private static readonly string[] Prefixes = { string.Empty, "Re: ", "Fwd: ", "RE: ", "FW: " };
-
-    private static readonly string[] Subjects =
-    {
-        "Meeting Tomorrow", "Project Update", "Quick Question", "Follow Up",
-        "Action Required", "Review Needed", "Status Report", "Weekly Summary",
-        "Important Notice", "Reminder", "Schedule Change", "Document Review",
-        "Contract Discussion", "Budget Update", "Team Meeting", "Client Call",
-    };
-
-    /// <summary>
-    /// Gets a random email subject.
-    /// </summary>
-    /// <param name="random">Random instance.</param>
-    /// <returns>Email subject.</returns>
-    public static string GetRandom(Random random)
-    {
-        var prefix = Prefixes[random.Next(Prefixes.Length)];
-        var subject = Subjects[random.Next(Subjects.Length)];
-        return prefix + subject;
-    }
-}
-
-/// <summary>
-/// Review notes generator.
-/// </summary>
-internal static class ReviewNotes
-{
-    private static readonly string[] Notes =
-    {
-        "Reviewed and marked responsive.",
-        "Contains privileged attorney-client communication.",
-        "Not relevant to the matter.",
-        "Contains trade secret information - needs redaction.",
-        "Duplicate of document DOC000001.",
-        "Foreign language content detected - needs translation.",
-        "Reviewed by reviewer on date.",
-        "Escalated for senior review.",
-        "Contains PII - redaction required.",
-        "Key document - flag for production.",
-    };
-
-    /// <summary>
-    /// Gets a random review note.
-    /// </summary>
-    /// <param name="random">Random instance.</param>
-    /// <returns>Review note.</returns>
-    public static string GetRandomNote(Random random)
-    {
-        return Notes[random.Next(Notes.Length)];
-    }
-}
-
-/// <summary>
-/// Encoding names.
-/// </summary>
-internal static class Encodings
-{
-    private static readonly string[] Values = { "UTF-8", "ASCII", "UTF-16", "ISO-8859-1", "Windows-1252" };
-
-    /// <summary>
-    /// Gets a random encoding name.
-    /// </summary>
-    /// <param name="random">Random instance.</param>
-    /// <returns>Encoding name.</returns>
-    public static string GetRandom(Random random) => Values[random.Next(Values.Length)];
-}
-
-/// <summary>
-/// Time zone names.
-/// </summary>
-internal static class TimeZones
-{
-    private static readonly string[] Values =
-    {
-        "UTC", "America/New_York", "America/Los_Angeles", "America/Chicago",
-        "Europe/London", "Europe/Paris", "Asia/Tokyo", "Asia/Shanghai",
-    };
-
-    /// <summary>
-    /// Gets a random time zone.
-    /// </summary>
-    /// <param name="random">Random instance.</param>
-    /// <returns>Time zone name.</returns>
-    public static string GetRandom(Random random) => Values[random.Next(Values.Length)];
-}
